functions/requirements.txt
firebase_functions~=0.1.0
functions-framework
requests
beautifulsoup4
google-cloud-firestore
firebase-admin>=6.0.0